+++ app/src/main/java/com/InnovaTechno/homemarket/Categories/Adapters/BAlcoliseesAdapter.java
@@ -1,17 +1,15 @@
-package com.InnovaTechno.homemarket.Categories.Adapter;
+package com.InnovaTechno.homemarket.Categories.Adapters;
 
 import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import android.widget.TextView;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.InnovaTechno.homemarket.Categories.Post.row;
+import com.InnovaTechno.homemarket.Categories.Models.row;
 import com.InnovaTechno.homemarket.R;
 
 import java.util.List;
