+++ app/src/main/java/com/InnovaTechno/homemarket/Categories/Adapters/ViandesAdapter.java
@@ -1,4 +1,4 @@
-package com.InnovaTechno.homemarket.Categories.Adapter;
+package com.InnovaTechno.homemarket.Categories.Adapters;
 
 import android.content.Context;
 import android.content.Intent;
@@ -13,7 +13,7 @@ import androidx.annotation.NonNull;
 import androidx.cardview.widget.CardView;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.InnovaTechno.homemarket.Categories.Post.PostPLaitiers;
+import com.InnovaTechno.homemarket.Categories.Models.PostViandes;
 import com.InnovaTechno.homemarket.Items_Detail.ItemDetails;
 import com.InnovaTechno.homemarket.R;
 import com.bumptech.glide.Glide;
@@ -21,12 +21,12 @@ import com.parse.ParseFile;
 
 import java.util.List;
 
-public class PLaitiersAdapter  extends RecyclerView.Adapter<PLaitiersAdapter.ViewHolder> {
+public class ViandesAdapter extends RecyclerView.Adapter<ViandesAdapter.ViewHolder> {
 
     private Context context;
-    private List<PostPLaitiers> posts;
+    private List<PostViandes> posts;
 
-    public PLaitiersAdapter(Context context, List<PostPLaitiers> posts) {
+    public ViandesAdapter(Context context, List<PostViandes> posts) {
         this.context = context;
         this.posts = posts;
     }
@@ -42,8 +42,8 @@ public class PLaitiersAdapter  extends RecyclerView.Adapter<PLaitiersAdapter.Vie
 
     @Override
     public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
-        PostPLaitiers postPLaitiers = posts.get(position);
-        holder.bind(postPLaitiers);
+        PostViandes postViandes = posts.get(position);
+        holder.bind(postViandes);
     }
 
     @Override
@@ -69,7 +69,7 @@ public class PLaitiersAdapter  extends RecyclerView.Adapter<PLaitiersAdapter.Vie
             tvPrice = itemView.findViewById(R.id.tvPrice);
 
         }
-        public void bind(PostPLaitiers post) {
+        public void bind(PostViandes post) {
             ParseFile image = post.getImage();
             if (image != null) {
                 Glide.with(context).load(image.getUrl()).into(ivFruits_Legumes);
