+++ app/src/main/java/com/InnovaTechno/homemarket/Categories/Boissons.java
@@ -14,8 +14,8 @@ import androidx.appcompat.widget.SearchView;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.InnovaTechno.homemarket.Categories.Adapter.BoissonsAdapter;
-import com.InnovaTechno.homemarket.Categories.Post.PostBoissons;
+import com.InnovaTechno.homemarket.Categories.Adapters.BoissonsAdapter;
+import com.InnovaTechno.homemarket.Categories.Models.PostBoissons;
 import com.InnovaTechno.homemarket.R;
 import com.InnovaTechno.homemarket.SignUpActivity;
 import com.parse.FindCallback;
