+++ app/src/main/java/com/InnovaTechno/homemarket/Categories/Models/PostAMenagers.java
@@ -1,18 +1,17 @@
-package com.InnovaTechno.homemarket.Categories.Post;
+package com.InnovaTechno.homemarket.Categories.Models;
 
 import com.parse.ParseClassName;
 import com.parse.ParseFile;
 import com.parse.ParseObject;
 
 
-@ParseClassName("Post")
-public class Post extends ParseObject {
+@ParseClassName("PostAMenagers")
+public class PostAMenagers extends ParseObject {
 
     public static final String KEY_NAME =  "name";
     public static final String KEY_IMAGE =  "productImage";
     public static final String KEY_DEVISE = "devise";
-    public static final String KEY_PRICE = "price2";
-    public static final String KEY_DESCRIPTION = "description";
+    public static final String KEY_PRICE = "price";
 
 
 
@@ -24,19 +23,13 @@ public class Post extends ParseObject {
         return getString(KEY_NAME);
     }
 
-    public String getDescription(){
-        return getString(KEY_DESCRIPTION);
+    public void setName(String name) {
+        put(KEY_NAME, name);
     }
 
     public ParseFile getImage (){
         return getParseFile(KEY_IMAGE);
     }
-
-    public void setName(String name) {
-        put(KEY_NAME, name);
-    }
-
-
 
     public void setImage (ParseFile parseFile){
         put(KEY_IMAGE, parseFile);
@@ -49,11 +42,4 @@ public class Post extends ParseObject {
     public void setDevise (String devise){
         put(KEY_DEVISE, devise);
     }
-
-    public void setDescription (String description){
-       put(KEY_DESCRIPTION, description);
-    }
-
-
-
 }
