+++ app/src/main/java/com/InnovaTechno/homemarket/Categories/Models/PostSucreries.java
@@ -1,25 +1,23 @@
-package com.InnovaTechno.homemarket.Categories.Post;
+package com.InnovaTechno.homemarket.Categories.Models;
 
 import com.parse.ParseClassName;
 import com.parse.ParseFile;
 import com.parse.ParseObject;
 
 
-@ParseClassName("PostBAlcolisees")
-public class PostBAlcolisees extends ParseObject {
+@ParseClassName("PostSucreries")
+public class PostSucreries extends ParseObject {
 
-    public static final String KEY_NAME =  "name";
-    public static final String KEY_IMAGE =  "productImage";
+    public static final String KEY_NAME = "name";
+    public static final String KEY_IMAGE = "productImage";
     public static final String KEY_DEVISE = "devise";
     public static final String KEY_PRICE = "price";
-
-
 
     public String getPrice() {
         return getString(KEY_PRICE);
     }
 
-    public String getName(){
+    public String getName() {
         return getString(KEY_NAME);
     }
 
@@ -27,19 +25,20 @@ public class PostBAlcolisees extends ParseObject {
         put(KEY_NAME, name);
     }
 
-    public ParseFile getImage (){
+    public ParseFile getImage() {
         return getParseFile(KEY_IMAGE);
     }
 
-    public void setImage (ParseFile parseFile){
+    public void setImage(ParseFile parseFile) {
         put(KEY_IMAGE, parseFile);
     }
 
-    public String getDevise () {
+    public String getDevise() {
         return getString(KEY_DEVISE);
     }
 
-    public void setDevise (String devise){
+    public void setDevise(String devise) {
         put(KEY_DEVISE, devise);
     }
+
 }
