+++ app/src/main/java/com/InnovaTechno/homemarket/Categories/Models/PostViandes.java
@@ -1,19 +1,17 @@
-package com.InnovaTechno.homemarket.Categories.Post;
+package com.InnovaTechno.homemarket.Categories.Models;
 
 import com.parse.ParseClassName;
 import com.parse.ParseFile;
 import com.parse.ParseObject;
 
 
-@ParseClassName("PostPCosmetiques")
-public class PostPCosmetiques extends ParseObject {
+@ParseClassName("PostViandes")
+public class PostViandes extends ParseObject {
 
     public static final String KEY_NAME =  "name";
     public static final String KEY_IMAGE =  "productImage";
     public static final String KEY_DEVISE = "devise";
     public static final String KEY_PRICE = "price";
-
-
 
     public String getPrice() {
         return getString(KEY_PRICE);
@@ -41,5 +39,8 @@ public class PostPCosmetiques extends ParseObject {
 
     public void setDevise (String devise){
         put(KEY_DEVISE, devise);
-    }
+
+
+}
+
 }
