+++ app/src/main/java/com/InnovaTechno/homemarket/Categories/Viandes.java
@@ -14,8 +14,8 @@ import androidx.appcompat.widget.SearchView;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.InnovaTechno.homemarket.Categories.Adapter.ViandesAdapter;
-import com.InnovaTechno.homemarket.Categories.Post.PostViandes;
+import com.InnovaTechno.homemarket.Categories.Adapters.ViandesAdapter;
+import com.InnovaTechno.homemarket.Categories.Models.PostViandes;
 import com.InnovaTechno.homemarket.R;
 import com.InnovaTechno.homemarket.SignUpActivity;
 import com.parse.FindCallback;
