+++ app/src/main/java/com/InnovaTechno/homemarket/Fragments/CartFragment.java
@@ -1,25 +1,20 @@
 package com.InnovaTechno.homemarket.Fragments;
 
-import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.InnovaTechno.homemarket.Categories.Adapter.PostAdapter;
-import com.InnovaTechno.homemarket.Categories.Post.Post;
+import com.InnovaTechno.homemarket.Categories.Models.Post;
 
 import com.InnovaTechno.homemarket.R;
 import com.InnovaTechno.homemarket.adapter.CartAdapter;
