+++ app/src/main/java/com/InnovaTechno/homemarket/adapter/CartAdapter.java
@@ -1,13 +1,10 @@
 package com.InnovaTechno.homemarket.adapter;
 
 import android.content.Context;
-import android.content.Intent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.BaseAdapter;
 import android.widget.Button;
-import android.widget.ImageSwitcher;
 import android.widget.ImageView;
 import android.widget.TextView;
 
@@ -15,8 +12,7 @@ import androidx.annotation.NonNull;
 import androidx.cardview.widget.CardView;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.InnovaTechno.homemarket.Cart;
-import com.InnovaTechno.homemarket.Categories.Post.Post;
+import com.InnovaTechno.homemarket.Categories.Models.Post;
 import com.InnovaTechno.homemarket.R;
 
 import java.util.List;
