+++ app/src/main/res/layout/cardview_categories.xml
@@ -8,7 +8,7 @@
     android:layout_marginBottom="10dp"
     android:foreground="?android:attr/selectableItemBackground"
     xmlns:cardview="http://schemas.android.com/apk/res-auto"
-    cardview:cardCornerRadius="5dp">
+    cardview:cardCornerRadius="2dp">
 
     <LinearLayout
         android:layout_width="match_parent"
